Move resource building methods to JSONAPI::Serializable
